Don't pretend R's RNG is thead-safe, it isn't! Use thread-local RNGs intead.
Also refactored the the simulation code to be as parallel as possible.
In particular, prepare the result vectors beforehand, and let each thread
fill in the results of its batch of lineages as it goes. This avoids having
to have a synchronization point after completing a time interval.

SANsimulatoR now depends on dqrng which provides the RNG, and on BH for
boost::random::binomial_distribution.