Marking Rcpp as firstprivate() in OpenMP parallel directives causes heap corruption
The steps vector was marked firstprivate(), which lead to heap corruption,
presumably because the IntegerVector copy constructor interacts with the GC
and firstprivate() caused it to be called from worker threads without proper
synchronization. The sympton were infinit hangs during function exit when
stack-allocated Rcpp objects were destructed.

Instead, we now pass an int* pointing to the vector elements instead.