Turned the repository inside-out to make it one huge R package
Before, the repository contained a shiny app in the top-level dir, and the
SANsimulatoR R package in a subdirectory. Now, it contains the R package (now
called SANjar) in the top-level directory, and a reduced version of the shiny
app in inst/SANModelExplorer. The parametersets are mostly oudated, and have
all been moved to parametersets-outdated for now.