Refactored MCMC code to allow more flexible initialization, reworked summarystats(), added sample.posterior() and sample.variables()
It's now possibly to easily sample the initial state of MCMC chains from the
posterior distribution of earlier MCMC runs, and to sample different parameters
from different sources.

The summarystats() function now fits into the overall MCMC framework of SANjar,
and was confirmed to return the same mode as previously for the LT47 MCMC results.